+++ lib/quick_travel/booking.rb
@@ -51,10 +51,7 @@ module QuickTravel
     #  :trailer => { :vehicle_type_id => nil , :length => nil  }
     # }
     def self.create(options = {})
-      default_options = { referral_code_id: DEFAULT_REFERRAL_CODE_ID }
-      options = { booking: default_options.merge(options) }
-
-      response = post_and_validate("#{api_base}.json", options)
+      response = post_and_validate("#{api_base}.json", booking: options)
       fail AdapterError.new(response) unless response['id']
 
       return nil unless response['id']